@@ -1,12 +1,12 @@
 plugins {
     id 'java'
-    id 'org.springframework.boot' version '2.7.14'
-    id 'io.spring.dependency-management' version '1.0.15.RELEASE'
+    id 'org.springframework.boot' version '3.1.4'
+    id 'io.spring.dependency-management' version '1.1.3'
 }
 
 ext.versions = [
-        springBootVersion                : '2.7.14',
-        springDependencyManagemnetVersion: '1.0.15.RELEASE',
+        springBootVersion                : '3.1.4',
+        springDependencyManagemnetVersion: '1.1.3',
         modelmapper                      : '3.1.1',
         springdocOpenAPI                 : '1.7.0',
         lombok                           : '1.18.26',
@@ -36,7 +36,7 @@ allprojects {
     version = "${versions.appVersion}"
 
     java {
-        sourceCompatibility = '11'
+        sourceCompatibility = '17'
     }
 
     repositories {
